[dreamc] modularize C code generation
